requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
flash==1.0.3
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
waitress==3.0.2
